requirements.txt
boto3>=1.34.0
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=0.27.0
